refactor: Standardize environment variable management
