+++ .gitignore
@@ -143,6 +143,7 @@ reminders.txt
 sorted_downloaded_file.csv
 backend/downloaded_file.csv
 google-chrome-stable_current_amd64.deb
+aircraftMapping.json
 
 # Debug
 debug/
