+++ backend/.env.example
@@ -2,14 +2,6 @@
 GO_ENV=development
 MYSQL_DSN="user:password@tcp(mysql:3306)/airaccidentdata?parseTime=true"
 
-# MySQL configuration
-MYSQL_ROOT_PASSWORD=
-MYSQL_PASSWORD=
-MYSQL_DATABASE=
-MYSQL_USER=
-MYSQL_PORT=
-MYSQL_HOST=mysql
-
 # Server configuration
 SERVER_ADDRESS=0.0.0.0:8080
