+++ backend/cmd/csvtomysql/main.go
@@ -12,21 +12,24 @@ import (
 	"net/http"
 	"net/url"
 	"os"
+	"path/filepath"
 	"sort"
 	"strconv"
+	"time"
 
+	"github.com/computers33333/airaccidentdata/internal/config"
 	"github.com/computers33333/airaccidentdata/internal/models"
-	"github.com/computers33333/airaccidentdata/internal/shared"
 	_ "github.com/go-sql-driver/mysql" // Blank identifier imports MySQL driver to initialize and register it.
+	"github.com/joho/godotenv"
 )
 
 // main is the entry point of the application, responsible for processing CSV data and inserting it into a MySQL database.
 func main() {
-	if err := shared.LoadEnv(); err != nil {
+	if err := loadEnv(); err != nil {
 		log.Fatalf("Error loading .env file: %v", err)
 	}
 
-	db, err := shared.SetupDatabase()
+	db, err := setupDatabase()
 	if err != nil {
 		log.Fatalf("Database setup failed: %v", err)
 	}
@@ -65,8 +68,8 @@ func processCSV(file *os.File, db *sql.DB) error {
 
 	// Sort records by ENTRY_DATE in descending order
 	sort.Slice(records, func(i, j int) bool {
-		entryDate1, err1 := shared.ParseDate(records[i][1])
-		entryDate2, err2 := shared.ParseDate(records[j][1])
+		entryDate1, err1 := ParseDate(records[i][1])
+		entryDate2, err2 := ParseDate(records[j][1])
 		if err1 != nil || err2 != nil {
 			return false
 		}
@@ -132,15 +135,15 @@ func parseRecordToIncident(record []string) (*models.Aircraft, *models.Accident,
 		AircraftOperator:   record[12],
 	}
 
-	entryDate, err := shared.ParseDate(record[1])
+	entryDate, err := ParseDate(record[1])
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("error parsing entry date: %v", err)
 	}
-	eventLocalDate, err := shared.ParseDate(record[2])
+	eventLocalDate, err := ParseDate(record[2])
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("error parsing event local date: %v", err)
 	}
-	eventLocalTime, err := shared.ParseTime(record[3])
+	eventLocalTime, err := ParseTime(record[3])
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("error parsing event local time: %v", err)
 	}
@@ -314,4 +317,74 @@ func getCoordinates(place string) (float64, float64, error) {
 	lng := geoResp.Results[0].Geometry.Location.Lng
 
 	return lat, lng, nil
+}
+
+// setupDatabase establishes a connection to the MySQL database.
+func setupDatabase() (*sql.DB, error) {
+	cfg := config.NewConfig()
+	dsn := cfg.DataSourceName
+
+	db, err := sql.Open("mysql", dsn)
+	if err != nil {
+		return nil, fmt.Errorf("could not open database: %w", err)
+	}
+
+	if err := db.Ping(); err != nil {
+		return nil, fmt.Errorf("database is not reachable: %w", err)
+	}
+
+	return db, nil
+}
+
+// loadEnv searches for the .env file starting in the current directory and moving up.
+func loadEnv() error {
+	dir, err := os.Getwd()
+	if err != nil {
+		return err
+	}
+
+	for {
+		if _, err := os.Stat(filepath.Join(dir, ".env")); err == nil {
+			return godotenv.Load(filepath.Join(dir, ".env"))
+		}
+
+		parentDir := filepath.Dir(dir)
+		if parentDir == dir {
+			return fmt.Errorf("root directory reached, .env file not found")
+		}
+		dir = parentDir
+	}
+}
+
+// atoiSafe converts string to int, returns 0 if conversion fails or the string is empty.
+func AtoiSafe(s string) int {
+	if s == "" {
+		return 0
+	}
+	value, err := strconv.Atoi(s)
+	if err != nil {
+		log.Printf("Error converting string to int: %v", err)
+		return 0
+	}
+	return value
+}
+
+// Helper function to parse a date string into time.Time, returns time.Time and error.
+func ParseDate(dateStr string) (time.Time, error) {
+	layout := "02-Jan-06"
+	t, err := time.Parse(layout, dateStr)
+	if err != nil {
+		return time.Time{}, fmt.Errorf("error parsing date '%s': %v", dateStr, err)
+	}
+	return t, nil
+}
+
+// Helper function to format a time string into time.Time, returns time.Time and error.
+func ParseTime(timeStr string) (string, error) {
+	layout := "15:04:05Z"
+	t, err := time.Parse(layout, timeStr)
+	if err != nil {
+		return "", fmt.Errorf("error parsing time '%s': %v", timeStr, err)
+	}
+	return t.Format("15:04:05"), nil
 }
